split up assemble tests
